fix: various search fixes
