+++ dadi/lib/model/index.js
@@ -994,7 +994,6 @@ Model.prototype.getIndexes = require('./collections/getIndexes')
 Model.prototype.getRevisions = require('./collections/getRevisions')
 Model.prototype.getStats = require('./collections/getStats')
 Model.prototype.getVersions = require('./collections/getVersions')
-Model.prototype.search = require('./collections/search')
 Model.prototype.update = require('./collections/update')
 
 module.exports = function (name, schema, connection, settings) {
